requirements.txt
asttokens==2.4.1
click==8.1.7
comm==0.2.2
geopandas==1.1.0
cloth-simulation-filter==1.1.4
Cython==3.0.9
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
ipykernel==6.29.3
ipython==8.22.2
jakteristics==0.6.0
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.6.1
jupyter_core==5.7.2
laspy==2.5.3
laszip==0.2.3
lazrs==0.5.3
lightgbm==4.3.0
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
numpy==1.26.4
packaging==24.0
parso==0.8.3
pexpect==4.9.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==23.12.0
Pygments==2.17.2
python-dateutil==2.9.0.post0
pyzmq==25.1.2
scikit-learn==1.4.1.post1
scipy==1.12.0
six==1.16.0
stack-data==0.6.3
threadpoolctl==3.4.0
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
typer==0.11.0
typing_extensions==4.10.0
wcwidth==0.2.13
xgboost==2.0.3


### install pdal via conda forge 
